allow changing commits to point back to previous test outputs
Not all test results will be changed by all output-changing commits. With this
change we keep track of which test cases still match previous golden outputs
and do not store another copy of that bitstream.  We will be able to generate
reports in the future describing which tests were changed by which commits and
how they changed (bitrate, ssim, and psnr anyway).