add support for keyword filter strings in output-changing-commits.txt
this is a way to declare which tests you expect to be changed by each commit
(and more importantly, which ones you do not expect to change).  If a commit
does not have a keyword filter it will take credit for any test case output
changes which do not match a keyword commit.  VBV tests will only ever update
their golden outputs if the 'vbv' keyword is declared by the commit