Fix: Fix version error thrown in performance test
Code Review: Mahesh

Details: For uhdkit, the version information is dumped in next row in csv so to read from
particular cell we are making changes but this is a temporary solution we need a right fix in uhdkit repo itself.
once we fix in uhdkit repo we will revert these changes.

Modified regeneratecsv function into two different sections based on encoder_binary_name.
For binary which are not x265 - regenerated csv, add 'Encoder version' column along with the other columns